requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
debugpy==1.8.1
Django==4.2.11
django-debug-toolbar==4.3.0
flake8==7.0.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
psycopg==3.1.18
psycopg-binary==3.1.18
pycodestyle==2.11.1
pyflakes==3.2.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
factory-boy==3.3.0 
